fixed problems during refreshing application
