+++ next.config.ts
@@ -1,7 +1,7 @@
 import type { NextConfig } from 'next';
 
 const nextConfig: NextConfig = {
-  reactStrictMode: false
+  reactStrictMode: false,
   /* config options here */
 };
