+++ src/app/styles/globals.scss
@@ -1,3 +1,3 @@
 @import 'tailwindcss';
 
-@import 'tailwind-variables';
+@import 'primereact-variables-for-tailwind';
