+++ src/app/styles/primereact-variables-for-tailwind.scss
